chore: score internally represented by i16 instead of i32
probably broke a lot of things

bench: 6334737

+++ toad/src/cli.rs
@@ -117,7 +117,7 @@ pub enum EngineCommand {
         square: Option<Square>,
 
         /// Evaluate `piece` at `square` with the provided endgame weight [0-100].
-        endgame_weight: Option<i32>,
+        endgame_weight: Option<i8>,
     },
 
     /// Performs a split perft on the current position at the supplied depth.
